[Shooting][Hotfix] pequeño arreglo
El daño debería de ser un float

+++ Assets/Scripts/GridSystem/BuildingSO.cs
@@ -54,7 +54,7 @@ public class BuildingSO : ScriptableObject
 
     public int width;
     public int heigth;
-    public int damage;
+    public float damage;
     public string name;
     public Transform prefab;
     public Transform visual;
